Split the web support to package dio_browser_adapter
