+++ dio/lib/browser.dart
@@ -1,2 +1,1 @@
-export 'src/adapters/browser_adapter.dart' show BrowserHttpClientAdapter;
 export 'src/dio/dio_for_browser.dart' show DioForBrowser;
