+++ dio/lib/src/dio_mixin.dart
@@ -15,7 +15,7 @@ import 'headers.dart';
 import 'interceptors/imply_content_type.dart';
 import 'options.dart';
 import 'progress_stream/io_progress_stream.dart'
-    if (dart.library.html) 'progress_stream/browser_progress_stream.dart';
+    if (dart.library.js_util) 'progress_stream/browser_progress_stream.dart';
 import 'response.dart';
 import 'response/response_stream_handler.dart';
 import 'transformer.dart';
