+++ dio/lib/src/multipart_file.dart
@@ -4,7 +4,7 @@ import 'dart:convert';
 import 'package:http_parser/http_parser.dart';
 
 import 'multipart_file/io_multipart_file.dart'
-    if (dart.library.html) 'multipart_file/browser_multipart_file.dart';
+    if (dart.library.js_util) 'multipart_file/browser_multipart_file.dart';
 import 'utils.dart';
 
 /// A file to be uploaded as part of a [MultipartRequest]. This doesn't need to
