+++ plugins/browser_adapter/lib/dio_browser_adapter.dart
@@ -1,0 +1,3 @@
+library dio_browser_adapter;
+
+export 'src/browser_adapter.dart';
